chore: refactor menu toggle
